requirements/dev.txt
-r ../requirements.txt

factory-boy==2.10.0
pylint==1.8.2
pylint-django==0.9.0
pytest==3.4.0
pytest-cov==2.5.1
pytest-django==3.1.2
snapshottest==0.5.0
pdbpp==0.9.1